+++ server/save.py
@@ -12,7 +12,7 @@ from config import SAVE_FILE
 from models import ALL_MODELS, Constants
 from models.db import db
 
-SAVE_VERSION = 18
+SAVE_VERSION = 19
 logger: logging.Logger = logging.getLogger("PlanarAllyServer")
 logger.setLevel(logging.INFO)
 
@@ -165,8 +165,12 @@ def upgrade(version):
 
         from models import GridLayer, Layer
 
-        db.execute_sql('CREATE TABLE IF NOT EXISTS "base_rect" ("shape_id" TEXT NOT NULL PRIMARY KEY, "width" REAL NOT NULL, "height" REAL NOT NULL, FOREIGN KEY ("shape_id") REFERENCES "shape" ("uuid") ON DELETE CASCADE)')
-        db.execute_sql('CREATE TABLE IF NOT EXISTS "shape_type" ("shape_id" TEXT NOT NULL PRIMARY KEY, FOREIGN KEY ("shape_id") REFERENCES "shape" ("uuid") ON DELETE CASCADE)')
+        db.execute_sql(
+            'CREATE TABLE IF NOT EXISTS "base_rect" ("shape_id" TEXT NOT NULL PRIMARY KEY, "width" REAL NOT NULL, "height" REAL NOT NULL, FOREIGN KEY ("shape_id") REFERENCES "shape" ("uuid") ON DELETE CASCADE)'
+        )
+        db.execute_sql(
+            'CREATE TABLE IF NOT EXISTS "shape_type" ("shape_id" TEXT NOT NULL PRIMARY KEY, FOREIGN KEY ("shape_id") REFERENCES "shape" ("uuid") ON DELETE CASCADE)'
+        )
 
         shape_types = [
             "asset_rect",
@@ -215,6 +219,7 @@ def upgrade(version):
         Constants.update(save_version=Constants.save_version + 1).execute()
     elif version == 15:
         from peewee import BooleanField
+
         migrator = SqliteMigrator(db)
         db.foreign_keys = False
         with db.atomic():
@@ -225,27 +230,45 @@ def upgrade(version):
         Constants.update(save_version=Constants.save_version + 1).execute()
     elif version == 16:
         from peewee import TextField
+
         migrator = SqliteMigrator(db)
         db.foreign_keys = False
         with db.atomic():
             migrate(
-                migrator.add_column("location", "unit_size_unit",
-                                    TextField(default="ft"))
+                migrator.add_column(
+                    "location", "unit_size_unit", TextField(default="ft")
+                )
             )
         db.foreign_keys = True
         Constants.update(save_version=Constants.save_version + 1).execute()
     elif version == 17:
         from peewee import BooleanField
+
         migrator = SqliteMigrator(db)
         db.foreign_keys = False
         with db.atomic():
             migrate(
-                migrator.add_column("polygon", "open_polygon", BooleanField(default=False)),
-                migrator.add_column("polygon", "line_width", IntegerField(default=2))
+                migrator.add_column(
+                    "polygon", "open_polygon", BooleanField(default=False)
+                ),
+                migrator.add_column("polygon", "line_width", IntegerField(default=2)),
             )
-            db.execute_sql("INSERT INTO polygon (shape_id, line_width, vertices, open_polygon) SELECT shape_id, line_width, points, 1 FROM multi_line")
-            db.execute_sql("DROP TABLE multi_line");
-            db.execute_sql("UPDATE shape SET type_ = 'polygon' WHERE type_ = 'multiline'")
+            db.execute_sql(
+                "INSERT INTO polygon (shape_id, line_width, vertices, open_polygon) SELECT shape_id, line_width, points, 1 FROM multi_line"
+            )
+            db.execute_sql("DROP TABLE multi_line")
+            db.execute_sql(
+                "UPDATE shape SET type_ = 'polygon' WHERE type_ = 'multiline'"
+            )
+        db.foreign_keys = True
+        Constants.update(save_version=Constants.save_version + 1).execute()
+    elif version == 18:
+        from peewee import TextField
+
+        migrator = SqliteMigrator(db)
+        db.foreign_keys = False
+        with db.atomic():
+            migrate(migrator.add_column("user", "email", TextField(null=True)))
         db.foreign_keys = True
         Constants.update(save_version=Constants.save_version + 1).execute()
     else: